refactor(dto): 移動 ApiResult 和 User 類別至 data 包，並刪除舊的 dto 包
